+++ pip_services_container/config/ContainerConfigReader.py
@@ -17,7 +17,7 @@ from .ContainerConfig import ContainerConfig
 class ContainerConfigReader(object):
 
     @staticmethod
-    def read_from_file(correlation_id, path):
+    def read_from_file(correlation_id, path, parameters):
         if path == None:
             raise ConfigException(correlation_id, "NO_PATH", "Missing config file path")
         
@@ -25,20 +25,20 @@ class ContainerConfigReader(object):
         ext = path[index + 1:].lower() if index > 0 else ''
 
         if ext == "json":
-            return ContainerConfigReader.read_from_json_file(correlation_id, path)
+            return ContainerConfigReader.read_from_json_file(correlation_id, path, parameters)
         elif ext == "yaml":
-            return ContainerConfigReader.read_from_yaml_file(correlation_id, path)
+            return ContainerConfigReader.read_from_yaml_file(correlation_id, path, parameters)
         
         # By default read as JSON
-        return ContainerConfigReader.read_from_json_file(correlation_id, path)
+        return ContainerConfigReader.read_from_json_file(correlation_id, path, parameters)
 
     @staticmethod
-    def read_from_json_file(correlation_id, path):
-        config = JsonConfigReader.read_config(correlation_id, path)
+    def read_from_json_file(correlation_id, path, parameters):
+        config = JsonConfigReader.read_config(correlation_id, path, parameters)
         return ContainerConfig.from_config(config)
 
     @staticmethod
-    def read_from_yaml_file(correlation_id, path):
-        config = YamlConfigReader.read_config(correlation_id, path)
+    def read_from_yaml_file(correlation_id, path, parameters):
+        config = YamlConfigReader.read_config(correlation_id, path, parameters)
         return ContainerConfig.from_config(config)